app_indoor/indoor-app/server/requirements.txt

Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
gevent==20.5.0
gevent-websocket==0.10.1

gunicorn==20.0.4
h5py==2.10.0
httplib2==0.9.2


SQLAlchemy==1.3.17
psycopg2-binary==2.8.4
Flask-Script==2.0.6
flask-sqlacodegen==1.1.7

websockets==8.1
PyJWT==1.7.1
Flask-Login=0.5.0 **
Flask-Bcrypt==0.7.1
